fix the communication problem by changing message type from double to float

+++ Initial_Material/controllers/localization_controller/localization_controller.c
@@ -50,7 +50,7 @@ static pose_t         _pose, _odo_acc, _odo_enc, _speed_enc;
 static pose_t         _pose_origin = {-2.9, 0, 0};
 double last_gps_time_s = 0.0f;
 double last_gps_send_time_s = 0.0f;
-double message[8] ;
+float message[8] ;
 static FILE *fp;
 static gsl_matrix*Cov_acc;
 static gsl_matrix*X_acc;
@@ -320,5 +320,10 @@ void send_mea() {
   message[5] = _odo_enc.y;
   message[6] = gsl_matrix_get(X_acc,0,0);
   message[7] = gsl_matrix_get(X_acc,1,0);
-  wb_emitter_send(emitter,message,8*sizeof(double)); 
+  // printf("emitter time: %g \n",time_now_s);
+  // printf("est: %g %g \n",message[0],message[1]);
+  // printf("est: %g %g \n",message[2],message[3]);
+  // printf("est: %g %g \n",message[4],message[5]);
+  // printf("est: %g %g \n",message[6],message[7]);
+  wb_emitter_send(emitter,message,8*sizeof(float)); 
 }
